# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tanzo API",
        "description": "API description",
        "contact": {
            "email": "utkarsh.wilson@regur.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get all projects of the logged in user",
                "description": "Get all projects of the logged in user",
                "operationId": "getAllProjects",
                "responses": {
                    "200": {
                        "description": "All projects fetched successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Error in fetching projects"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "description": "Create a new project",
                "operationId": "createProject",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "user_id": {
                                                "type": "int64"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Project name",
                                    "description": "Project description",
                                    "user_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Error in creating project"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/projects/{projectId}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get a project by ID",
                "description": "Get a project by ID",
                "operationId": "getOneProject",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of Project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All projects details fetched successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Error in fetching user details"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update a project",
                "description": "Update a project",
                "operationId": "updateProject",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of Project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "user_id": {
                                                "type": "int64"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "New Project name",
                                    "description": "New Project description",
                                    "user_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project updated successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Error in updating project"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Delete a project",
                "description": "Delete a project",
                "operationId": "deleteProject",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of Project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project deleted successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Error in deleting project"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization"
            }
        }
    }
}